+++ examples/booking_app/test/feature_hotel_details/view/hotel_details_golden_test.dart
@@ -1,25 +1,20 @@
 import '../../helpers/golden_helper.dart';
-import '../../helpers/models/scenario.dart';
 import '../../stubs.dart';
 import '../factory/hotel_details_factory.dart';
 
 void main() {
   runGoldenTests([
-    generateDeviceBuilder(
+    buildScenario(
       widget: hotelDetailsFactory(
         hotel: Stub.hotel1,
       ), //example: Stubs.emptyList
-      scenario: Scenario(
-        name: 'hotel_details_loading',
-      ),
+      scenario: 'hotel_details_loading',
     ),
-    generateDeviceBuilder(
+    buildScenario(
       widget: hotelDetailsFactory(
         hotel: Stub.hotel1Loaded,
       ), //example: Stubs.emptyList
-      scenario: Scenario(
-        name: 'hotel_details_loaded',
-      ),
+      scenario: 'hotel_details_loaded',
     ),
   ]);
 }