+++ examples/booking_app/test/feature_hotel_search/view/hotel_search_golden_test.dart
@@ -1,33 +1,33 @@
 import '../../helpers/golden_helper.dart';
-import '../../helpers/models/scenario.dart';
 import '../../stubs.dart';
 import '../factory/hotel_search_factory.dart';
 
 void main() {
   runGoldenTests([
-    generateDeviceBuilder(
-      scenario: Scenario(name: 'hotel_search_empty'),
+    buildScenario(
+      scenario: 'hotel_search_empty',
       widget: hotelSearchFactory(
         hotels: Stub.paginatedListEmpty,
       ), //example: Stubs.emptyList
     ),
-    generateDeviceBuilder(
-      scenario: Scenario(name: 'hotel_search_success'),
+    buildScenario(
+      scenario: 'hotel_search_success',
       widget: hotelSearchFactory(
         hotels: Stub.paginatedListHotelThree,
         capacityFilterData: Stub.capacityFilterDataTwoPersonsOneRoom,
         dateRangeFilterData: Stub.dateRangeFilterData,
       ), //example:  Stubs.success
     ),
-    generateDeviceBuilder(
-        widget: hotelSearchFactory(
-          hotels: Stub.paginatedListEmptyLoading,
-        ), //loading
-        scenario: Scenario(name: 'hotel_search_loading')),
-    generateDeviceBuilder(
+    buildScenario(
+      widget: hotelSearchFactory(
+        hotels: Stub.paginatedListEmptyLoading,
+      ), //loading
+      scenario: 'hotel_search_loading',
+    ),
+    buildScenario(
         widget: hotelSearchFactory(
           hotels: Stub.paginatedListError,
         ),
-        scenario: Scenario(name: 'hotel_search_error'))
+        scenario: 'hotel_search_error')
   ]);
 }