+++ examples/booking_app/test/flutter_test_config.dart
@@ -1,30 +1,56 @@
 import 'dart:async';
 
+import 'package:alchemist/alchemist.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:golden_toolkit/golden_toolkit.dart';
-//ignore: depend_on_referenced_packages
-import 'package:visibility_detector/visibility_detector.dart';
 
-import 'helpers/booking_app_file_comparator.dart';
+import 'helpers/goldens_file_comparator.dart';
+
+/// Flag indicating if the tests are running in a CI environment
+/// Note: The environment variable may differ based on the platform.
+const bool _isRunningInCi = bool.fromEnvironment('CI');
+
+/// Resolves the file path for the golden image based on the name and environment
+FutureOr<String> _filePathResolver(String name, String env) {
+  // Resolve the theme name by removing the theme from name and placing it
+  // in the correct directory
+  if (name.endsWith('_light')) {
+    name = 'light_theme/${name.replaceAll('_light', '')}';
+  } else if (name.endsWith('_dark')) {
+    name = 'dark_theme/${name.replaceAll('_dark', '')}';
+  }
+
+  final fileName = 'goldens/${env.toLowerCase()}/$name.png';
+  goldenFileComparator = GoldensFileComparator(fileName);
+  return fileName;
+}
 
 Future<void> testExecutable(FutureOr<void> Function() testMain) async =>
-    GoldenToolkit.runWithConfiguration(
-      () async {
+    AlchemistConfig.runWithConfig(
+      config: AlchemistConfig(
+        goldenTestTheme: GoldenTestTheme(
+          backgroundColor: Colors.grey,
+          borderColor: Colors.transparent,
+          nameTextStyle: const TextStyle(
+            color: Colors.black,
+            fontSize: 16,
+          ),
+        ),
+        platformGoldensConfig: const PlatformGoldensConfig(
+          enabled: !_isRunningInCi,
+          obscureText: false,
+          renderShadows: true,
+          filePathResolver: _filePathResolver,
+        ),
+        ciGoldensConfig: const CiGoldensConfig(
+          enabled: _isRunningInCi,
+          obscureText: false,
+          renderShadows: true,
+          filePathResolver: _filePathResolver,
+        ),
+      ),
+      run: () async {
         WidgetsApp.debugAllowBannerOverride = false;
-        VisibilityDetectorController.instance.updateInterval = Duration.zero;
-
-        await loadAppFonts();
         return testMain();
       },
-      config: GoldenToolkitConfiguration(
-        enableRealShadows: true,
-        fileNameFactory: (String name) {
-          final fileName = 'goldens/$name.png';
-
-          goldenFileComparator = BookingAppFileComparator(fileName);
-
-          return fileName;
-        },
-      ),
     );